[refactor]#108 유저 행위 기록, 분석 interface 분리
