+++ src/main/java/org/netway/dongnehankki/analytics/presentation/AnalyticsController.java
@@ -6,13 +6,13 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.netway.dongnehankki.analytics.application.AnalyticsService;
+import org.netway.dongnehankki.analytics.application.AnalyticsServiceImpl;
 import org.netway.dongnehankki.analytics.dto.response.WeeklyHourlyStatsResponse;
 import org.netway.dongnehankki.global.common.ApiResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @Tag(name = "통계", description = "통계 관련 API")
