+++ src/main/java/org/netway/dongnehankki/store/application/StoreService.java
@@ -3,7 +3,7 @@ package org.netway.dongnehankki.store.application;
 
 import java.util.Collections;
 import java.util.List;
-import org.netway.dongnehankki.analytics.application.AnalyticsService;
+import org.netway.dongnehankki.analytics.application.AnalyticsServiceImpl;
 import org.netway.dongnehankki.analytics.domain.ActivityType;
 import org.netway.dongnehankki.post.domain.Post;
 import org.netway.dongnehankki.post.repository.PostRepository;
@@ -40,7 +40,7 @@ public class StoreService {
 	private final MenuRepository menuRepository;
 	private final UserRepository userRepository;
 	private final PostRepository postRepository;
-	private final AnalyticsService analyticsService;
+	private final AnalyticsServiceImpl analyticsService;
 
 	@Transactional
 	public StoreResponse getStoreById(Long storeId, Long loginId) {
